Migrate Impress parser to it's own package to prevent auto-run
Spring package scanning picks up all components in the same package
or below in the package tree.  We want to run the IMPReSS parser
separately, so this migrates the impress parser to it's own package
and cleans up some of the leftover spring bean definitions from the
previous version of the IMPReSS parser.