Fix icon grid bug
For Categorical data, if either of the sex specific (one sex only) tests are
significant, but the combined sexes test is not significant, we need to check
each sex independently.  If either of the sexes are significant by themselves,
a phenotype call will have been made for that sex, even though the overall
(combined) result is not significant.